Add test cases

+++ src/main/java/com/example/order/order/models/Item.java
@@ -10,7 +10,7 @@ public class Item {
     private String name;
 
 
-    private float price;
+    private double price;
 
     @Override
     public String toString() {
@@ -23,7 +23,7 @@ public class Item {
     public Item() {
 
     }
-    public Item(String name, float price ) {
+    public Item(String name, double price ) {
         this.name = name;
         this.price = price;
     }
@@ -36,11 +36,11 @@ public class Item {
         this.name = name;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public void setPrice(float price) {
+    public void setPrice(double price) {
         this.price = price;
     }
 }
